fix: uint128 for gat

+++ packages/core/contracts/data-verification-mechanism/implementation/ResultComputationV2.sol
@@ -54,7 +54,7 @@ library ResultComputationV2 {
      */
     function getResolvedPrice(
         Data storage data,
-        uint256 minTotalVotes,
+        uint128 minTotalVotes,
         uint256 minModalVotes
     ) internal view returns (bool isResolved, int256 price) {
         if (data.totalVotes > minTotalVotes && data.voteFrequency[data.currentMode] > minModalVotes) {
